pyproject.toml
[build-system]
requires = [ "setuptools>=60", "setuptools-scm>=8" ]

[tool.setuptools]
packages = [
    "bris_fiab", 
    "bris_fiab.anemoi_plugins.inference.downscale", 
    "bris_fiab.checkpoint",
    "bris_fiab.orography",
]

[project]
name = "bris-fiab"
version = "0.1.0"
description = "Running the bris model in Forecast-in-a-Box"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anemoi-datasets==0.5.26",
    "anemoi-graphs==0.6.2",
    "anemoi-inference==0.7.0",
    "anemoi-models==0.4.1",
    "anemoi-transform==0.1.14",
    "anemoi-utils==0.4.29",
    "earthkit-data==0.16.4",
    "ecmwf-api-client==1.6.5",
    "metpy>=1.7.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "rioxarray>=0.19.0",
    "scipy>=1.16.1",
    "trimesh>=4.7.3",
    "pyyaml>=6.0.1",
    "metpy>=1.7.1",
]

[project.entry-points."anemoi.inference.inputs"]
downscaled_mars = "bris_fiab.anemoi_plugins.inference.downscale:DownscaledMarsInput"
adiabatically_corrected_mars = "bris_fiab.anemoi_plugins.inference.apply_adiabatic_corrections:AdiabaticallyCorrectedMarsInput"

# [project.entry-points."anemoi.inference.pre_processors"]
# downscale = "bris_fiab.anemoi_plugins.inference.downscale:DownscalePreProcessor"
# apply_adiabatic_corrections = "bris_fiab.anemoi_plugins.inference.apply_adiabatic_corrections:AdiabaticCorrectionPreProcessor"
